benchmark/ML workflows/mobilenet/mobilenet/requirements.txt
onnxruntime
numpy
Pillow
onnx
line-profiler